chore : update README for next steps
We will now try to interact with zoom on the api side, and get a list of existing zoom events.
These existing events will first be manually created on the zoom interface.
There does not seem to exist a supported zoom SDK for node.js applications, so we will directly interact with
zoom's HTTP api.