feat: Add Route to get All Meetings
On Router, add Route to get all meetings.
On Zoom Adapter, implement port function to get all zoom meetings.
DTO used is the same as the object we get from the Zoom API for simplicity.
If we get a 401 error when trying to access the zoom api, we reset the Oauth token flow.
Indeed, the access token only lasts one hour, so it will eventually time out.
We should implement some kind of retry strategy in such a case ; for now, we only reset the token so that
on the next Zoom API call, we load a new Oauth Token again. This is not perfect but will work for now.