feat: add react-query library to web interface
To interact with api, use react-query library.
Use meetings endpoint to get a list of all existing meetings.
We can see here that we need to duplicate the typing files (Meeting.ts) on the front and back projects.
With more time, I would extract those to a dedicated package if the need arises strongly.