feat: Display overlays for existing meetings
Logic is as follows : For each day, get associated existing meetings (using sameDay fn).
For each of these meetings, then, display a component whose height corresponds to the length of the meeting.
The component is positioned as to fit the start/end hours of its related meeting.
When clicking on this component, show the meeting start/end times and zoom join url.
I used some generic components for the modals, as we mentioned the UI did not really matter.