feat: add api route to create a new Meeting
On express server, add route to create a new meeting.
On the boundary side, implement logic to create a meeting on Zoom through its api.
Same logic as before on token failure.
To be noted : on the POST request sent to the Zoom API, a lot of parameters are not injected currently, but written directly inside the code : meeting owner, meeting timezone, meeting password, etc..
The only injected parameters right now are the meeting date and duration, and topic.
This is for ease of development, and given the given timeframe. With more time, these parameters should also be injected through the api endpoint, or default business parameters etc..