feat: disable hours in the past
Set hours in the past as "disabled", for now simply a UI effect.
We will prevent drag-and-dropping on past hours, and hence creating events in the past.
We do this after having played with the zoom api : creating events in the past does not seem possible
with the zoom api (if the start time of the meeting is in the past, the api fallbacks to the current request time).