feat: first version of drag and drop functionality, to create a new meeting
Using mouse event listeners (onMouseMove, onMouseDown etc..), first version of drag and drop feature.
Feature renders quite laggy, to be investigated, and code is messy : logic should be extracted in a dedicated hook. To refactor.
The positioning of the event being created is done absolutely w.r. to the height of the page and calendar. By calculating the total screen height and calendar height, we can calculate, when dragging the newly creating component corresponding to a new meeting, how many hours it will last, thanks to its height + start Y position. we know the event's start hour by calculating its offset w.r. to the calendar height, and by knowing the default height of a single hour tile (rule of three).
Right now, new events can only be created with steps of 1 hour ; when creating an event between hour tiles, we made the opiniated choice of rounding down the start hour to the closest hour below, and rounding up the end hour to the closest hour above.